regulation_task/requirements.txt
numpy==1.24.3
sentence-transformers==2.2.2
huggingface_hub==0.13.3
PyMuPDF==1.22.0
Pillow==9.0.0
pytesseract==0.3.10
anthropic>=0.5.0
gradio
